backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain==0.3.7
langchain-huggingface
langchain-community==0.3.7
chromadb==0.4.22
python-dotenv==1.0.1
pydantic==2.10.2
transformers==4.40.0
torch==2.4.0
sentence-transformers==2.7.0
